Moved views implementations to presentation layer
