+++ app/src/main/java/me/grechka/yamblz/yamblzweatherapp/presentation/AboutFragment.java
@@ -1,4 +1,4 @@
-package me.grechka.yamblz.yamblzweatherapp;
+package me.grechka.yamblz.yamblzweatherapp.presentation;
 
 
 import android.os.Bundle;
@@ -10,6 +10,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
+
+import me.grechka.yamblz.yamblzweatherapp.R;
 
 
 /**
