+++ app/src/main/java/me/grechka/yamblz/yamblzweatherapp/presentation/activity/MainView.java
@@ -1,13 +1,10 @@
-package me.grechka.yamblz.yamblzweatherapp.activity;
+package me.grechka.yamblz.yamblzweatherapp.presentation.activity;
 
-import com.arellomobile.mvp.InjectViewState;
 import com.arellomobile.mvp.MvpView;
 import com.arellomobile.mvp.viewstate.strategy.AddToEndSingleStrategy;
 import com.arellomobile.mvp.viewstate.strategy.SingleStateStrategy;
 import com.arellomobile.mvp.viewstate.strategy.SkipStrategy;
 import com.arellomobile.mvp.viewstate.strategy.StateStrategyType;
-
-import dalvik.annotation.TestTargetClass;
 
 /**
  * Created by Grechka on 15.07.2017.
