+++ app/src/main/java/me/grechka/yamblz/yamblzweatherapp/presentation/settings/SettingsView.java
@@ -1,0 +1,11 @@
+package me.grechka.yamblz.yamblzweatherapp.presentation.settings;
+
+import com.arellomobile.mvp.MvpView;
+
+/**
+ * Created by Grechka on 15.07.2017.
+ */
+
+public interface SettingsView extends MvpView {
+
+}
