--- app/src/main/java/me/grechka/yamblz/yamblzweatherapp/settings/SettingsView.java
@@ -1,13 +1,0 @@
-package me.grechka.yamblz.yamblzweatherapp.settings;
-
-import com.arellomobile.mvp.MvpView;
-import com.arellomobile.mvp.viewstate.strategy.AddToEndSingleStrategy;
-import com.arellomobile.mvp.viewstate.strategy.StateStrategyType;
-
-/**
- * Created by Grechka on 15.07.2017.
- */
-
-public interface SettingsView extends MvpView {
-
-}
