+++ app/src/main/res/layout/activity_main.xml
@@ -4,5 +4,5 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="me.grechka.yamblz.yamblzweatherapp.activity.MainActivity">
+    tools:context=".presentation.activity.MainActivity">
 </FrameLayout>
