+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="me.grechka.yamblz.yamblzweatherapp.activity.MainActivity">
+    tools:context=".presentation.activity.MainActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
