+++ app/src/main/res/layout/screen_weather.xml
@@ -4,7 +4,7 @@
     android:id="@+id/swiperefresh"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="me.grechka.yamblz.yamblzweatherapp.activity.MainActivity">
+    tools:context=".presentation.activity.MainActivity">
 
     <android.support.constraint.ConstraintLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
